Move list item component to separate file
